Organise packages in parser test folder
