+++ src/test/java/seedu/address/logic/parser/reminder/DeleteReminderCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.reminder;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,7 +8,6 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_REMINDER;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.reminder.DeleteReminderCommand;
-import seedu.address.logic.parser.reminder.DeleteReminderCommandParser;
 
 /**
  * As we are only doing white-box testing, our test cases do not cover path variations
